fix weird drawing caused by unsiged ints

+++ components/engine/include/rasterizer.h
@@ -10,8 +10,8 @@ typedef uint16_t pixel_t;
 
 struct bitmap {
 	pixel_t *buf;
-	size_t size_x;
-	size_t size_y;
+	int size_x;
+	int size_y;
 };
 
 pixel_t *bitmap_get_coord(struct bitmap *bmp, size_t x, size_t y);
@@ -22,7 +22,7 @@ struct vec world_to_screen_coords(struct vec screen_pos,
 /**
  * Draw a rect of `color` into `bmp`. `x`, `y` are in screen coords.
  */
-void rast_fillrect(struct bitmap *bmp, uint32_t x, uint32_t y, uint32_t w,
-				   uint32_t h, pixel_t color);
+void rast_fillrect(struct bitmap *bmp, int x, int y, int w,
+				   int h, pixel_t color);
 
 #endif // __RASTERIZER_H__
